requirements.txt
fastapi
uvicorn
faiss-cpu
langchain_community
langchain_openai
langchain_core
python-dotenv